Modified: Split benchmark_runner.cpp and pattern_benchmark/execution.cpp to smaller files as they were getting long.
